+++ src/main/java/com/sportinggoods/commands/Command.java
@@ -1,5 +1,9 @@
 package com.sportinggoods.commands;
 
+/**
+ * Functional interface representing a command.
+ */
+@FunctionalInterface
 public interface Command {
     void execute();
 }
