--- src/main/java/com/sportinggoods/controller/SportingGoodsMain2.java
@@ -1,4 +1,0 @@
-package com.sportinggoods.controller;
-
-public class SportingGoodsMain2 {
-}
